+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [1.0.5] - 2019-03-18
+
+### Added
+
+- Admins can view Payment details, including a link to the relevant transaction on paypal.com.
+
+### Changed
+
+- Only Payments which have been executed will be visible throughout the client to avoid confusion.
+- Fixed bug in ContactInformationContainer that caused the app to crash when no data had been entered for Contact Info.
+
 ## [1.0.4] - 2019-02-28
 
 ### Added